moved class to own folder
